requirements.txt
asgiref==3.5.0
boto3==1.20.54
botocore==1.23.54
certifi==2021.10.8
charset-normalizer==2.0.11
dj-database-url==0.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-decouple==2.1
django-extensions==3.1.5
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.12.3
gunicorn==20.1.0
idna==3.3
jmespath==0.10.0
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
psycopg2-pool==1.1
python-dateutil==2.8.2
python-decouple==3.6
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==5.3.0
